feat: Hide main window when opening financial objective
This commit adds a line of code to hide the main window when the financial objective is opened in the MainUiController. This improves the user experience by removing unnecessary clutter and focusing the user's attention on the financial objective.

Refactor risk_ui_controller.py to use a list instead of a dictionary for the RiskProfilingController's on_save method. This ensures that the results are saved in the correct order and improves code readability.

Update RiskProfilingUI's save_clicked signal to emit a list instead of a dictionary. This aligns with the changes made in the RiskProfilingController.

Delete main.ui file as it is no longer needed and has been replaced by other UI files.

Recent commits:
- feat(*): add new risk ui
- chore: Add result.txt to .gitignore
- refactor(main.py): improve code readability and add new result file creation
- feat: Improve Financial Objective UI options with numbers
- refactor(file_utils): add methods for managing result.txt file
- refactor(main_ui_controller): update button click event handler
- feat: Add static utilities methods
- feat: Add Financial Objective UI controller and view
- chore(*): remove useless files